requirements.txt
ansible-core
dnspython
